portfolio/requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.6
defusedxml==0.7.1
django-allauth
django-bootstrap
django-bootstrap-v5==1.0.2
django-bootstrap4==3.0.0
django-contact-form
django-crispy-forms
django-injector
django-package-monitor
django-rest-framework
django-rest-swagger
django-suit2
django-watson
djangorestframework
entrypoints
idna
injector==0.18.4
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
ng2django==20.7.13
notebook==6.3.0
numpy==1.20.2
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-sql==1.2.1
python3-openid==3.2.0
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
requirements-parser==0.1.0
scikit-learn==0.24.1
scipy==1.6.2
semantic-version==2.8.5
Send2Trash==1.5.0
sentry-sdk==1.0.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
